Remove vendor directory
Vendoring in Go is not generally safe for libraries to use.

In this particular case, vendoring grpc results in vendoring
x/net/trace, which has a non-idempotent `init()` function, which panics
on invocations after the first. The result is that importers of this
package which also import x/net/trace will panic on init.